requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
PyYAML>=6.0
tabulate>=0.9.0
schedule>=1.2.0
